requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
waitress==3.0.2
beautifulsoup4==4.12.3
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
